Remove dashes from .py file name
Dashes can be used in .py files but this only complicates things. If you want to
use .py file as a module and import it doing:

from file-name-with-dashes.py import *

Python will complain with 'syntax error'. To import it, it requires special
syntax:

module_name = __import__('file-name-with-dashes')

But this is unnecessarily complicated.